refactor: Enum 타입 Entity 필드 default -> @Enumerated(EnumType.STRING) 변경
