+++ 003 Code/backend/club-service/club-service/src/main/java/com/wcd/clubservice/entity/TagEntity.java
@@ -1,4 +1,4 @@
-package com.wcd.clubservice.jpa;
+package com.wcd.clubservice.entity;
 
 import jakarta.persistence.*;
 import lombok.Data;
