+++ 003 Code/backend/club-service/club-service/src/main/java/com/wcd/clubservice/repository/ClubMemberRepository.java
@@ -1,5 +1,7 @@
-package com.wcd.clubservice.jpa;
+package com.wcd.clubservice.repository;
 
+import com.wcd.clubservice.entity.ClubEntity;
+import com.wcd.clubservice.entity.ClubMemberEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
