+++ 003 Code/backend/club-service/club-service/src/main/java/com/wcd/clubservice/repository/ClubRepository.java
@@ -1,5 +1,6 @@
-package com.wcd.clubservice.jpa;
+package com.wcd.clubservice.repository;
 
+import com.wcd.clubservice.entity.ClubEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
