+++ 003 Code/backend/club-service/club-service/src/main/java/com/wcd/clubservice/repository/TagRepository.java
@@ -1,5 +1,6 @@
-package com.wcd.clubservice.jpa;
+package com.wcd.clubservice.repository;
 
+import com.wcd.clubservice.entity.TagEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TagRepository extends JpaRepository<TagEntity, Long> {
