+++ 003 Code/backend/user-service/user-service/src/main/java/com/wcd/userservice/entity/UserEntity.java
@@ -33,6 +33,7 @@ public class UserEntity {
     @Column(nullable = false)
     private LocalDate birthday;
 
+    @Enumerated(EnumType.STRING)
     @Column(nullable = false)
     private Gender gender;
