+++ 003 Code/backend/user-service/user-service/src/main/java/com/wcd/userservice/service/UserService.java
@@ -4,8 +4,7 @@ import com.wcd.userservice.dto.RegenerateTokenDto;
 import com.wcd.userservice.dto.TokenDto;
 import com.wcd.userservice.dto.UserDto;
 import com.wcd.userservice.dto.UserEvaluationDto;
-import com.wcd.userservice.vo.RequestUpdateUser;
-import org.springframework.security.core.userdetails.UserDetailsService;
+import com.wcd.userservice.vo.request.RequestUpdateUser;
 
 // Authentication에 인증하기 위한 자격으로 들어가려면 UserDetailService를 상속받아 구현해줘야함
 public interface UserService{
