+++ 003 Code/backend/user-service/user-service/src/main/java/com/wcd/userservice/service/UserServiceImpl.java
@@ -11,8 +11,8 @@ import com.wcd.userservice.exception.CustomException;
 import com.wcd.userservice.repository.UserEvaluationRepository;
 import com.wcd.userservice.repository.UserRepository;
 import com.wcd.userservice.security.jwt.JwtTokenProvider;
-import com.wcd.userservice.vo.RequestUpdateUser;
-import com.wcd.userservice.vo.ResponseClub;
+import com.wcd.userservice.vo.request.RequestUpdateUser;
+import com.wcd.userservice.vo.response.ResponseClub;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.convention.MatchingStrategies;
