+++ 003 Code/backend/user-service/user-service/src/main/java/com/wcd/userservice/vo/request/RequestUser.java
@@ -1,4 +1,4 @@
-package com.wcd.userservice.vo;
+package com.wcd.userservice.vo.request;
 
 import com.wcd.userservice.enums.Gender;
 import jakarta.validation.constraints.NotNull;
